backend/pyproject.toml
[project]
name = "askdocs-backend"
version = "0.1.0"
description = "AskDocs backend — FastAPI + Postgres/pgvector"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.32",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.30",
  "alembic>=1.13",
  "pgvector>=0.3",
  "pydantic>=2.9",
  "pydantic-settings>=2.6",
  "email-validator>=2.2",
  "python-jose[cryptography]>=3.3",
  "bcrypt>=4.1",
  "aiofiles>=24.1",
  "python-multipart>=0.0.12",
  "pypdf>=5.0",
  "python-docx>=1.1",
  "langchain-text-splitters>=0.3",
  "tiktoken>=0.8",
  "openai>=1.50",
  "httpx>=0.27",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.24",
  "ruff>=0.7",
  "mypy>=1.13",
]
fallback-parser = [
  "unstructured[pdf]>=0.15",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
